pyproject.toml
[tool.poetry]
name = "vcpilotbackend"
version = "0.1.0"
description = "Data entry assistant for VC"
authors = ["Dieter Annys <dieter.annys@persgroep.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
openai = "^1.12.0"
tavily-python = "^0.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
